refactor: modularize form logic
