+++ assets/js/index.js
@@ -1,0 +1,2 @@
+import './form-utils.js';
+import './form-chains.js';
