+++ index.html
@@ -757,7 +757,7 @@
 <script src="/node_modules/bootstrap/dist/js/bootstrap.bundle.js"></script>
 <script src="/assets/js/settings.js?v=0.0.5"></script>
 <script src="/assets/js/auth.js?v=0.0.5"></script>
-<script src="/assets/js/form-and-events.js?v=0.0.5"></script>
+<script type="module" src="/assets/js/index.js"></script>
 <script src="/assets/js/main.js?v=0.0.5"></script>
 <script src="/assets/js/auth.js?v=0.0.5"></script>
 <script src="/assets/js/arketype-wallets.js?v=0.0.5"></script>
