+++ pages/nonfungibles.html
@@ -288,7 +288,7 @@
 <script src="/node_modules/bootstrap/dist/js/bootstrap.bundle.js"></script>
 <script src="/assets/js/settings.js"></script>
 <script src="/assets/js/plugins.js"></script>
-<script src="/assets/js/form-and-events.js"></script>
+<script type="module" src="/assets/js/index.js"></script>
 <script src="/assets/js/main.js"></script>
 <script src="/assets/js/auth.js"></script>
 <script src="/assets/js/arketype-nonfungibles.js"></script>
